% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmash.R
\name{mnmash}
\alias{mnmash}
\title{multivariate, multiple regression extension of Adaptive SHrinkage (ASH)}
\usage{
mnmash(X, Y, U, omega, pi_0 = NULL, control = NULL)
}
\arguments{
\item{X}{[N, P] X data matrix}

\item{Y}{[N, J] Y data matrix}

\item{U}{[K, J, J] prior matrices}

\item{omega}{[L, 1] grids for priors}

\item{pi_0}{[K * L, 1] weights for priors}

\item{control}{\{tol = 1E-5, maxiter = 10000, logfile = NULL, n_cpu = 1\} list of runtime variables}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...
}
\author{
Gao Wang and Wei Wang
}
\references{
...
}

